makefile: introduce the Nbuild system
What is Nbuild?

Nbuild is a new build system designed to simplify and streamline
the build process for the megarepo. Just simply make a Makefile
in the project and then append the folder contaning the Makefile
to the local Nbuild file and it will magically be included in the
build process.

Changes made:
- Added Nbuild system to main makefile
- Removed old build scripts and marked them as depricated
- Updated .gitignore to include .vscode/settings.json
- Added colour and icon settings to Nbuild files for VSCode
- Created example Nbuild files in the usr/ directory as a starting point

Tested-by: 0x4248 <60709927+0x4248@users.noreply.github.com>
Signed-off-by: 0x4248 <60709927+0x4248@users.noreply.github.com>